requirements.txt
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-appconf==1.0.3
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-croppie==1.3
django-filter==2.1.0
django-recaptcha==2.0.4
django-rest-swagger==2.2.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
django-tinymce4-lite==1.7.4
django-uuslug==1.1.8
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djoser==1.7.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jsmin==2.2.2
Markdown==3.1.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==5.4.1
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.1
webencodings==0.5.1
